Refactor duplicated service utilities.
