+++ op-service/client/timeout.go
@@ -1,0 +1,8 @@
+package client
+
+import (
+	"time"
+)
+
+// DefaultDialTimeout is a default timeout for dialing a client.
+const DefaultDialTimeout = 5 * time.Second
